update verfier service
