+++ config/default_config.go
@@ -1,7 +1,7 @@
 package config
 
 const DefaultRPC = "https://rpc-nebulas-testnet.u2u.xyz"
-const DefaultSubnetAppStoreAddr = "0x984a3633256eD119EA118B8Ccf1dCe5D0FBc626E"
-const DefaultSubnetProviderAddr = "0x55F67461A9aA953eFA0D6645AD3d3Ac272733bAF"
-const DefaultSubnetProviderUptimeAddr = "0x530e2BB5ceFA2ffBcB0F615B16491D79Dc3dde80"
-const DefaultSubnetVerifier = "0x20869fEDD2e6a25FB447fb202F66a65E8512df3D"
+const DefaultSubnetAppStoreAddr = "0xaa0Dce508602d0eF4E1501F4C0eA6cBDD3b8018F"
+const DefaultSubnetProviderAddr = "0x3D99fd0A0423a0d3E6808d7726b12f4854B57C6D"
+const DefaultSubnetProviderUptimeAddr = "0x73e871C41E2Cc5c568C294e14EC5F8E005174d25"
+const DefaultSubnetVerifier = "0x837Cd3C2E6780dd22A9eaFb75f656B3349CC7050"
