+++ core/apps/registry.go
@@ -3,24 +3,26 @@ package apps
 import (
 	"context"
 	"fmt"
-	"math/big"
 
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/common"
+	atypes "github.com/unicornultrafoundation/subnet-node/core/apps/types"
 )
 
-func (s *Service) ReportUsage(ctx context.Context, appId, usedCpu, usedGpu, usedMemory, usedStorage, usedUploadBytes, usedDownloadBytes, duration *big.Int, signature []byte) (common.Hash, error) {
-	providerId := big.NewInt(s.accountService.ProviderID())
+func (s *Service) ReportUsage(ctx context.Context, usage *atypes.ResourceUsage, signature []byte) (common.Hash, error) {
 	// Create a new transactor
 	key, err := s.accountService.NewKeyedTransactor()
 	if err != nil {
 		return common.Hash{}, err
 	}
 
-	peerId := s.peerId.String()
-
 	// Call the ReportUsage function from the ABI
-	tx, err := s.accountService.AppStore().ReportUsage(key, providerId, appId, peerId, usedCpu, usedGpu, usedMemory, usedStorage, usedUploadBytes, usedDownloadBytes, duration, signature)
+	tx, err := s.accountService.AppStore().ReportUsage(
+		key, usage.AppId, usage.ProviderId, usage.PeerId,
+		usage.UsedCpu, usage.UsedGpu, usage.UsedMemory, usage.UsedStorage,
+		usage.UsedUploadBytes, usage.UsedDownloadBytes, usage.Duration,
+		usage.Timestamp,
+		signature)
 	if err != nil {
 		return common.Hash{}, err
 	}
